Split navigation into top-level package and environment.unknown_space
Move reusable navigation primitives (pathing, constants, occupancy grid
mixin) to railroad.navigation with zero environment imports. Move
UnknownSpaceEnvironment and its helpers (frontiers, laser, mapping) to
railroad.environment.unknown_space. This resolves the structural coupling
between navigation utilities and environment-specific code.

Also move ActiveSkill protocol from skill/protocols.py into environment.py
to eliminate the intra-package circular import, allowing all __init__.py
files to use eager imports.
