+++ packages/railroad/src/railroad/environment/environment.py
@@ -4,12 +4,43 @@ import itertools
 import math
 import warnings
 from abc import ABC, abstractmethod
-from typing import Callable, Collection, Dict, List, Optional, Set, Tuple
+from typing import Callable, Collection, Dict, List, Optional, Protocol, Set, Tuple, runtime_checkable
 
 from railroad._bindings import Action, Fluent, GroundedEffect, State
 from railroad.core import Operator
 
-from .skill import ActiveSkill
+
+@runtime_checkable
+class ActiveSkill(Protocol):
+    """Protocol for tracking execution of a single action."""
+
+    @property
+    def is_done(self) -> bool:
+        """Whether the skill has completed."""
+        ...
+
+    @property
+    def is_interruptible(self) -> bool:
+        """Whether this skill can be interrupted mid-execution."""
+        ...
+
+    @property
+    def upcoming_effects(self) -> List[Tuple[float, GroundedEffect]]:
+        """Effects still to occur, with expected times."""
+        ...
+
+    @property
+    def time_to_next_event(self) -> float:
+        """Time until next effect. May block in physical mode."""
+        ...
+
+    def advance(self, time: float, env: "Environment") -> None:
+        """Advance to given time, apply due effects to environment."""
+        ...
+
+    def interrupt(self, env: "Environment") -> None:
+        """Interrupt this skill, applying partial effects to environment."""
+        ...
 
 
 class Environment(ABC):
