+++ packages/railroad/src/railroad/environment/skill/protocols.py
@@ -2,47 +2,11 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, List, Protocol, Tuple, runtime_checkable
+from typing import Protocol, runtime_checkable
 
 import numpy as np
 
-from railroad._bindings import GroundedEffect
-
-if TYPE_CHECKING:
-    from ..environment import Environment
-
-
-@runtime_checkable
-class ActiveSkill(Protocol):
-    """Protocol for tracking execution of a single action."""
-
-    @property
-    def is_done(self) -> bool:
-        """Whether the skill has completed."""
-        ...
-
-    @property
-    def is_interruptible(self) -> bool:
-        """Whether this skill can be interrupted mid-execution."""
-        ...
-
-    @property
-    def upcoming_effects(self) -> List[Tuple[float, GroundedEffect]]:
-        """Effects still to occur, with expected times."""
-        ...
-
-    @property
-    def time_to_next_event(self) -> float:
-        """Time until next effect. May block in physical mode."""
-        ...
-
-    def advance(self, time: float, env: "Environment") -> None:
-        """Advance to given time, apply due effects to environment."""
-        ...
-
-    def interrupt(self, env: "Environment") -> None:
-        """Interrupt this skill, applying partial effects to environment."""
-        ...
+from ..environment import ActiveSkill
 
 
 @runtime_checkable
