+++ packages/railroad/src/railroad/environment/unknown_space/__init__.py
@@ -1,0 +1,12 @@
+"""UnknownSpaceEnvironment and helpers for frontier-based exploration."""
+
+from .environment import UnknownSpaceEnvironment
+from .types import Frontier, NavigationConfig
+from ..types import Pose
+
+__all__ = [
+    "Frontier",
+    "NavigationConfig",
+    "Pose",
+    "UnknownSpaceEnvironment",
+]
