+++ packages/railroad/src/railroad/navigation/occupancy_grid_mixin.py
